Updating #5 from develop

+++ src/main/java/ISPP/G5/INDVELOPERS/controllers/GameController.java
@@ -19,8 +19,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import ISPP.G5.INDVELOPERS.models.Developer;
 import ISPP.G5.INDVELOPERS.models.Game;
-import ISPP.G5.INDVELOPERS.repositories.GameRepository;
 import ISPP.G5.INDVELOPERS.services.DeveloperService;
 import ISPP.G5.INDVELOPERS.services.GameService;
 
@@ -53,11 +53,21 @@ public class GameController {
 	
 	@PostMapping("/add")
 	public ResponseEntity<String> addGame(@RequestBody Game game) throws NotFoundException {
+		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+		UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+		Developer developer = this.developerService.findByUsername(userDetails.getUsername());
+		boolean isPremium = false;
+		if(developer.getIsPremium() != null) {
+			isPremium = developer.getIsPremium();
+		}
 		try {
-			Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-			UserDetails userDetails = (UserDetails) authentication.getPrincipal();
-			game.setCreator(this.developerService.findByUsername(userDetails.getUsername()));
-			return ResponseEntity.status(HttpStatus.CREATED).body(this.gameService.addGame(game));
+			if(this.gameService.findAll().stream().anyMatch(g -> g.getTitle().equals(game.getTitle())))
+				throw new IllegalArgumentException("There's alredy a game with that title");
+			if(isPremium == false && game.getPrice() != 0.0) 
+				throw new IllegalArgumentException("Only premium developers can sell non-free games");
+			if(isPremium == false && (this.gameService.findByMyGames(developer.getId()).size() + 1 == 6))
+				throw new IllegalArgumentException("Non premium developers only can have a maximun of five games published");
+			return ResponseEntity.status(HttpStatus.CREATED).body(this.gameService.addGame(game, developer));
 		} catch(IllegalArgumentException e) {
 			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
 		}
@@ -65,15 +75,24 @@ public class GameController {
 	
 	@PutMapping("/edit/{id}")
 	public ResponseEntity<String> updateGame(@PathVariable String id, @RequestBody Game game) throws NotFoundException{
+		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+		UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+		Developer developer = developerService.findByUsername(userDetails.getUsername());
 		Game gameData = this.gameService.findById(id);
+		List<Game> allGames = this.gameService.findAll();
+		allGames.remove(gameData);
 		try {
+			if(allGames.stream().anyMatch(g -> g.getTitle().equals(game.getTitle())))
+				throw new IllegalArgumentException("There's alredy a game with that title");
+			if (!gameData.getCreator().getId().equals(developer.getId())) 
+				throw new IllegalArgumentException("Only the creator of the game can edit it");
 			gameData.setTitle(game.getTitle());
 			gameData.setDescription(game.getDescription());
 			gameData.setRequirements(game.getRequirements());
 			gameData.setPrice(game.getPrice());
 			gameData.setIsNotMalware(game.getIsNotMalware());
 			gameData.setIdCloud(game.getIdCloud());
-			 return new ResponseEntity<>(this.gameService.updateGame(gameData), HttpStatus.OK);
+			return new ResponseEntity<>(this.gameService.updateGame(gameData), HttpStatus.OK);
 		} catch(IllegalArgumentException e){
 			return new ResponseEntity<String>(HttpStatus.BAD_REQUEST);
 		}
@@ -81,11 +100,18 @@ public class GameController {
 	
 	@DeleteMapping("/delete/{id}")
 	public ResponseEntity<HttpStatus> deleteGameById(@PathVariable("id") String id) throws NotFoundException{
+		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+		UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+		Developer developer = this.developerService.findByUsername(userDetails.getUsername());
+		Game game = this.gameService.findById(id);
 		try {
+			if (!game.getCreator().getId().equals(developer.getId())) { 
+				throw new IllegalArgumentException("Only the creator of the game can remove it");
+			}
 			this.gameService.deleteGame(id);
 			return new ResponseEntity<>(HttpStatus.NO_CONTENT);
 		} catch(IllegalArgumentException e) {
-			return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+			return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
 		}
 	}
 	
@@ -93,16 +119,17 @@ public class GameController {
 	public ResponseEntity<List<Game>> getGameByTitle(@PathVariable String title) {
 		try {
 			return ResponseEntity.ok(this.gameService.findByTitle(title));
-		} catch(IllegalArgumentException | NotFoundException e) {
+		} catch(IllegalArgumentException e) {
 			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
 		}
 	}
 	
 	@GetMapping("/findByDeveloper/{developerUsername}")
-	public ResponseEntity<List<Game>> getGameByDeveloper(@PathVariable String developerUsername) {
+	public ResponseEntity<List<Game>> getGameByDeveloper(@PathVariable String developerUsername) throws NotFoundException {
 		try {
-			return ResponseEntity.ok(this.gameService.findByDeveloper(developerUsername));
-		} catch(IllegalArgumentException | NotFoundException e) {
+			Developer developer = this.developerService.findByUsername(developerUsername);
+			return ResponseEntity.ok(this.gameService.findByDeveloper(developer.getId()));
+		} catch(IllegalArgumentException e) {
 			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
 		}
 	}
@@ -110,17 +137,20 @@ public class GameController {
 	@GetMapping("/findMyGames")
 	public ResponseEntity<List<Game>> getGameByMyGames() {
 		try {
-			return ResponseEntity.ok(this.gameService.findByMyGames());
+			Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+			UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+			Developer developer = this.developerService.findByUsername(userDetails.getUsername());
+			return ResponseEntity.ok(this.gameService.findByMyGames(developer.getId()));
 		} catch(IllegalArgumentException | NotFoundException e) {
 			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
 		}
 	}
 	
 	@GetMapping("/{id}")
-	public ResponseEntity<Game> getGameById(@PathVariable String id) {
+	public ResponseEntity<Game> getGameById(@PathVariable String id) throws NotFoundException {
 		try {
 			return ResponseEntity.ok(this.gameService.findById(id));
-		} catch(IllegalArgumentException | NotFoundException e) {
+		} catch(IllegalArgumentException e) {
 			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
 		}
 	}
